+++ src/modules/bookstore/useCases/DeleteLibrary/DeleteLibraryUseCase.ts
@@ -1,3 +1,5 @@
+import { UsersRepository } from "@modules/accounts/infra/repositories/implementations/UsersRepository";
+import { IUsersRepository } from "@modules/accounts/infra/repositories/IUsersRepository";
 import { DeleteLibraryDTO } from "@modules/bookstore/@types";
 import { ILibraryRepository } from "@modules/bookstore/infra/repositories/ILibraryRepository";
 import { LibraryRepository } from "@modules/bookstore/infra/repositories/implementations/LibraryRepository";
@@ -8,16 +10,21 @@ import { inject, injectable } from "tsyringe";
 class DeleteLibraryUseCase {
   constructor(
     @inject(LibraryRepository)
-    private libraryRepository: ILibraryRepository
+    private libraryRepository: ILibraryRepository,
+    @inject(UsersRepository)
+    private usersRepository: IUsersRepository
   ) {}
 
-  async execute({ libraryId }: DeleteLibraryDTO): Promise<void> {
-    const libraryExists = this.libraryRepository.findById({ libraryId });
+  async execute({ libraryId, userId }: DeleteLibraryDTO): Promise<void> {
+    const userExists = await this.usersRepository.findById({ userId });
+    if (userExists.admin === false)
+      throw new AppError("User is not admin", 404);
 
+    const libraryExists = this.libraryRepository.findById({ libraryId });
     if (!libraryExists)
       throw new AppError("This book doesn't exists! Try again", 404);
 
-    await this.libraryRepository.delete({ libraryId });
+    await this.libraryRepository.delete({ libraryId, userId });
   }
 }
